Fixing maxvcpu mismatch with topology issue, making it dynamic!
Maxvcpu is hardcoded for the specific testcase in virsh_emulatorpin.py.
When smp value, vcpu_cores, vcpu_threads, vcpu_sockets is specified at the
time of running the testcase then the hardcoded value will fail.
Adding the code to get this value dynamically based on the smp value
provided in the test suite config file.
1 example is that  hardcoded value is 7 in virsh_emulatorpin.cfg but in the testsuite
config smp value given is 32 not topology will be affected in this case my check
will help.
Signed-off-by: Anushree-Mathur <anushree.mathur@linux.ibm.com>